Add notifications feature with enhanced localization and user experience
- Add metaData to notification
- Updated routing to include a dedicated notifications path.
