+++ app/ui/components/notificationBell.scss
@@ -1,99 +1,100 @@
 @import "../../variables.module.scss";
 
 .notification-bell {
-    margin-right: 15px;
-    display: flex;
-    align-items: center;
-    
-    .bell-icon {
-      font-size: 20px;
-      cursor: pointer;
-      color: rgba(0, 0, 0, 0.65);
-      
-      &:hover {
-        color: #fc8282;
-      }
+  margin-right: 15px;
+  display: flex;
+  align-items: center;
+
+  .bell-icon {
+    font-size: 20px;
+    cursor: pointer;
+    color: rgba(0, 0, 0, 0.65);
+
+    &:hover {
+      color: #fc8282;
     }
   }
-  
-  .notification-popover {
+}
+
+.ant-popover {
+  &.notification-popover {
     width: 350px;
-    
+
     .ant-popover-inner {
       padding: 0;
     }
-  }
-  
-  .notification-popup {
-    width: 100%;
-    max-height: 500px;
-    display: flex;
-    flex-direction: column;
-    
-    .notification-header {
-      padding: 12px 16px;
-      border-bottom: 1px solid #f0f0f0;
+
+    .notification-popup {
+      width: 100%;
+      max-height: 90vh;
       display: flex;
-      justify-content: space-between;
-      align-items: center;
-      
-      h5 {
-        margin: 0;
-      }
-    }
-    
-    .notification-content {
-      max-height: 350px;
-      overflow-y: auto;
-      padding: 0;
-      
-      .notification-loading {
-        display: flex;
-        justify-content: center;
-        padding: 24px;
-      }
-      
-      .notification-item {
+      flex-direction: column;
+
+      .notification-header {
         padding: 12px 16px;
-        cursor: pointer;
-        transition: background-color 0.3s;
-        
-        &:hover {
-          background-color: #f5f5f5;
+        border-bottom: 1px solid #f0f0f0;
+        display: flex;
+        justify-content: space-between;
+        align-items: center;
+
+        h5 {
+          margin: 0;
         }
-        
-        &.unread {
-          background-color: #e6f7ff;
-          
-          &:hover {
-            background-color: #bae7ff;
-          }
-        }
-        
-        .notification-item-content {
+      }
+
+      .notification-content {
+        max-height: 350px;
+        overflow-y: auto;
+        padding: 0;
+
+        .notification-loading {
           display: flex;
-          flex-direction: column;
-          width: 100%;
-          
-          .ant-typography {
-            margin-bottom: 4px;
-            
-            &:last-child {
-              margin-bottom: 0;
+          justify-content: center;
+          padding: 24px;
+        }
+
+        .notification-item {
+          padding: 12px 16px;
+          cursor: pointer;
+          transition: background-color 0.3s;
+
+          &:hover {
+            background-color: #f5f5f5;
+          }
+
+          &.unread {
+            background-color: #e6f7ff;
+
+            &:hover {
+              background-color: #bae7ff;
             }
           }
-          
-          .notification-time {
-            font-size: 12px;
+
+          .notification-item-content {
+            display: flex;
+            flex-direction: column;
+            width: 100%;
+
+            .ant-typography {
+              margin-bottom: 4px;
+
+              &:last-child {
+                margin-bottom: 0;
+              }
+            }
+
+            .notification-time {
+              font-size: 12px;
+            }
           }
         }
       }
-    }
-    
-    .notification-footer {
-      padding: 12px 16px;
-      border-top: 1px solid #f0f0f0;
-      text-align: center;
+
+      .notification-footer {
+        padding: 12px 16px;
+        border-top: 1px solid #f0f0f0;
+        text-align: center;
+      }
     }
   }
-  
+}
