+++ i18n/locales/vi.json
@@ -531,7 +531,7 @@
   "Notification": {
     "title": "Thông báo",
     "title_page": "Thông báo của bạn",
-    "description": "Thông báo của website",
+    "description": "Thông báo của website Tiểu Phương Crochet",
     "error": {
       "title": "Lỗi",
       "fetch": "Không thể tải thông báo. Vui lòng thử lại sau!",
